+++ spec/support/pages/admin_dashboard_page.rb
@@ -118,14 +118,6 @@ class EditModal < PageFragment
     find('.cancel-link').click
     AdminDashboardPage.new(context: context)
   end
-
-  def reset_password
-    find('.reset-password-link').click
-  end
-
-  def reset_password_status
-    find('.reset-password .success')
-  end
 end
 
 class EditJournalFragment < PageFragment