+++ CHANGELOG.md
@@ -16,6 +16,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 ### Removed
 - Appeal button
 - Figure caption (temporary)
+- Password reset buttons
 
 ### Fixed
 -